tdx-signed-fit-image: override signing parameters from TI
At the time of this writing, the TI layer meta-ti-bsp is setting some of
the FIT signing parameters in a machine configuration file by using a
bitbake "?=" assignment which ends up overriding the openembedded and
meta-toradex-security defaults. By itself this is an issue, but in
addition the layer also sets FIT_SIGN_ALG="rsa4096" without making
FIT_SIGN_NUMBITS="4096" which causes build errors.

As a temporary solution to the above problem we override their defaults
by using override "k3". Once fixes to the layer are upstreamed the
present commit should be safely removed.

Related-to: TOR-3556

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>