tdx-signed-fit-image: override signing parameters from TI
At the time of this writing, the TI layer meta-ti-bsp is setting some of
the FIT signing parameters in a machine configuration file which ends up
overriding the openembedded and meta-toradex-security defaults. In
addition to this the layer also sets FIT_SIGN_ALG="rsa4096" without
making FIT_SIGN_NUMBITS="4096" which causes build errors. To solve this,
we here override their defaults by using override "k3". This in turn has
a side effect of requiring users to use an override if they want to
change the defaults we set; this is something we will "solve" via
documentation.

Related-to: TOR-3556

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>