u-boot: add patches to detect closed k3-based device
Add new patches to U-Boot applicable to k3-based devices (currently
verdin-am62) with the goal to allow the device state (open / closed)
detection. This functionality is in turn utilized by the hardening
features such as the "command whitelisting" and the "CLI access
prevention" to provide different behavior in each state.

Related-to: TOR-3556

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>