ti-k3-secdev-secboot.inc: remove stamp file for the do_unpack task
This is an attempt to solve a build failure happening on Jenkins.
Basically we remove the stamp file for the task causing it to run even
when it's deemed up-to-date. The task will still undergo sstate
optimizations, so this stamp removal is not expected to cause
unnecessary builds to occur.

Related-to: TOR-3840

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>